requirements.txt
numpy==1.13.3
torchvision==0.2.0
pandas==0.22.0
pillow==5.0.0
tqdm